backend/fastapi-nginx-gunicorn/requirement.txt
fastapi==0.109.0
gunicorn==20.1.0
uvicorn[standard]==0.20.0
pymongo==4.6.1
python-dotenv==1.0.1
certifi==2023.11.17